Move error boundary to same control
