+++ packages/react-examples/src/react-charting/DeclarativeChart/DeclarativeChart.Basic.Example.tsx
@@ -2,7 +2,35 @@ import * as React from 'react';
 import { Dropdown, IDropdownOption } from '@fluentui/react/lib/Dropdown';
 import { TextField, ITextFieldStyles } from '@fluentui/react/lib/TextField';
 import { DeclarativeChart, DeclarativeChartProps, IDeclarativeChart, Schema } from '@fluentui/react-charting';
-import ErrorBoundary from './ErrorBoundary';
+
+interface IErrorBoundaryProps {
+  children: React.ReactNode;
+}
+
+interface IErrorBoundaryState {
+  hasError: boolean;
+  error: string;
+}
+
+class ErrorBoundary extends React.Component<IErrorBoundaryProps, IErrorBoundaryState> {
+  public static getDerivedStateFromError(error: Error) {
+    // Update state so the next render will show the fallback UI.
+    return { hasError: true, error: `${error.message} ${error.stack}` };
+  }
+
+  constructor(props: IErrorBoundaryProps) {
+    super(props);
+    this.state = { hasError: false, error: '' };
+  }
+
+  public render() {
+    if (this.state.hasError) {
+      return <h1>${this.state.error}</h1>;
+    }
+
+    return this.props.children;
+  }
+}
 
 interface IDeclarativeChartState {
   selectedChoice: string;
