--- packages/react-examples/src/react-charting/DeclarativeChart/ErrorBoundary.tsx
@@ -1,32 +1,0 @@
-import * as React from 'react';
-
-interface IErrorBoundaryProps {
-  children: React.ReactNode;
-}
-
-interface IErrorBoundaryState {
-  hasError: boolean;
-  error: string;
-}
-
-class ErrorBoundary extends React.Component<IErrorBoundaryProps, IErrorBoundaryState> {
-  public static getDerivedStateFromError(error: Error) {
-    // Update state so the next render will show the fallback UI.
-    return { hasError: true, error: `${error.message} ${error.stack}` };
-  }
-
-  constructor(props: IErrorBoundaryProps) {
-    super(props);
-    this.state = { hasError: false, error: '' };
-  }
-
-  public render() {
-    if (this.state.hasError) {
-      return <h1>${this.state.error}</h1>;
-    }
-
-    return this.props.children;
-  }
-}
-
-export default ErrorBoundary;
